新增功能：区域管理、系统日志和用户管理页面
- 实现了 Regions.vue，用于管理区域数据，并具备搜索、添加、编辑和删除功能。

- 创建了 SystemLogs.vue，用于显示系统日志，并具备搜索和分页功能。

- 开发了 Users.vue，用于用户管理，包括用户创建、编辑、密码重置和删除功能。

- 为 Regions、SystemLogs 和 Users 组件中的表单添加了验证规则。

- 为所有组件集成了 CRUD 操作的 API 调用。

+++ BackEnd/HailongConsulting.API/Repositories/VisitStatisticRepository.cs
@@ -58,7 +58,7 @@ public class VisitStatisticRepository : Repository<VisitStatistic>, IVisitStatis
             .ToListAsync();
     }
 
-    public async Task<IEnumerable<(string PageUrl, string? PageTitle, uint TotalViews)>> GetTopPagesAsync(int topCount, int days)
+    public async Task<IEnumerable<(string PageUrl, string? PageTitle, int TotalViews)>> GetTopPagesAsync(int topCount, int days)
     {
         var startDate = DateOnly.FromDateTime(DateTime.Now.AddDays(-days));
         
@@ -75,15 +75,15 @@ public class VisitStatisticRepository : Repository<VisitStatistic>, IVisitStatis
             .Take(topCount)
             .ToListAsync();
 
-        return result.Select(r => (r.PageUrl ?? "", r.PageTitle, (uint)r.TotalViews));
+        return result.Select(r => (r.PageUrl ?? "", r.PageTitle, (int)r.TotalViews));
     }
 
-    public async Task<uint> GetPageTotalViewsAsync(string pageUrl)
+    public async Task<int> GetPageTotalViewsAsync(string pageUrl)
     {
         var totalViews = await _dbSet
             .Where(v => v.PageUrl == pageUrl && v.IsDeleted == 0)
-            .SumAsync(v => (long)v.VisitCount);
+            .SumAsync(v => v.VisitCount);
         
-        return (uint)totalViews;
+        return (int)totalViews;
     }
 }
